Extract workspace local environment APIs
